calc правки

+++ lw8/calc/include/Calculator.h
@@ -10,8 +10,8 @@ public:
 	static std::string ProcessCommand(const std::string& command);
 
 private:
-	static std::vector<long long> ParseNumbers(const std::string& input);
-	static std::string Calculate(char operation, const std::vector<long long>& nums);
+	static std::vector<int> ParseNumbers(const std::string& input);
+	static std::string Calculate(char operation, const std::vector<int>& nums);
 };
 
 #endif // CALC_CALCULATOR_H
